chore: fixed landing home page
