+++ crowdin/messages.json
@@ -3045,8 +3045,6 @@
   "-207591979": "Deriv <0>multipliers</0> combines the upside of leverage trading without the downside of losing more than your stake. This means that when the market moves in your favour, you'll multiply your potential profits. If the market moves against your prediction, your losses are limited only to your stake. To find out more, visit the <0>Multipliers</0> page.",
   "-771604768": "How do I trade multipliers?",
   "-1755209362": "Watch <0>this video</0> to know more about trading multipliers.",
-  "-240686434": "Reliable.",
-  "-396178446": "100+ tradeable assets",
   "-456233007": "It is a partnership arrangement where a payment agent is authorised to process deposits and withdrawals for our clients.",
   "-898245410": "Are there any fees I need to pay to become a payment agent on Deriv?",
   "-173947736": "As a payment agent, you determine your commission per transaction, subject to our established thresholds.",
@@ -3089,8 +3087,8 @@
   "-1793883644": "Trade FX and CFDs on a customisable, easy-to-use trading platform.",
   "-1781346745": "Discover Deriv X now",
   "-279398477": "Trade the world’s markets on <0>Binary.com</0>’s classic platform.",
-  "-1781881664": "Check out our other platform",
   "-2130720993": "Check out our other platforms",
+  "-1781881664": "Check out our other platform",
   "-1825471709": "A whole new trading experience on a powerful yet easy to use platform.",
   "-204498543": "Trade with leverage and tight spreads for better returns on successful trades.",
   "-2115275974": "CFDs",
@@ -3585,11 +3583,13 @@
   "-1859212587": "Frequently asked questions - Payment Agent",
   "-1618901770": "Help centre | Frequently asked questions | Trading | Deriv",
   "-2085755369": "Frequently asked questions - Trading",
-  "-542307428": "world map",
+  "-240686434": "Reliable.",
+  "-396178446": "100+ tradeable assets",
   "-1894673375": "mobile app deriv go",
   "-505163259": "laptop dtrader",
   "-1633415532": "laptop deriv x",
   "-404595524": "laptop dmt5 eu",
+  "-542307428": "world map",
   "-313388618": "Dtrader trading platform at Deriv",
   "-642822922": "What our clients say about Deriv",
   "-97177023": "Trade 24/7 on forex, stocks, synthetic indices, and commodities",
