--- src/pages/home/_data.ts
@@ -1,22 +1,0 @@
-import { TContent, THeaderItem } from './_types'
-
-export const header_items: THeaderItem[] = [
-    { id: 0, text: '_t_Simple._t_' },
-    { id: 1, text: '_t_Flexible._t_' },
-    { id: 2, text: '_t_Reliable._t_' },
-]
-
-export const contents: TContent[] = [
-    { id: 0, text: '_t_Tight spreads_t_' },
-    { id: 1, text: '_t_Sharp prices_t_' },
-    { id: 2, text: '_t_24x7 trading_t_' },
-    { id: 3, text: '_t_100+ tradeable assets_t_' },
-    { id: 4, text: '_t_20+ years of experience_t_' },
-]
-
-export const contents_ppc: TContent[] = [
-    { id: 3, text: '_t_Tight spreads_t_' },
-    { id: 2, text: '_t_Sharp prices_t_' },
-    { id: 1, text: '_t_100+ tradeable assets_t_' },
-    { id: 0, text: '_t_20+ years of experience_t_' },
-]
